basic functioning

+++ dairy-management/src/main/java/com/dairy/dairy_management/entity/Product.java
@@ -13,6 +13,5 @@ public class Product {
 
     private String name;
     private String unit;
-
-    private BigDecimal pricePerUnit;
+    private  double pricePerUnit;
 }
